extract useOriginDrawer
